pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]
name = "nomad-tfsc-general"
description = "General nomad classes for thin film solar cells"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Michael Götte", email = "michael.goette@helmholtz-berlin.de" },
]
maintainers = [
    { name = "Michael Götte", email = "michael.goette@helmholtz-berlin.de" },
]
license = { file = "LICENSE" }
dependencies = [
    "nomad-lab>=1.3.16",
    'nomad-baseclasses @ git+https://github.com/nomad-hzb/nomad-baseclasses.git',
]

[project.urls]
Repository = "https://github.com/nomad-hzb/nomad-tfsc-general"

[project.optional-dependencies]
dev = ["ruff", "pytest", "structlog"]

[tool.uv]
extra-index-url = [
  "https://gitlab.mpcdf.mpg.de/api/v4/projects/2187/packages/pypi/simple",
]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "utils",
]

# Same as Black.
line-length = 110
indent-width = 4

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # isort
    "I",
    # pylint 
    "PL",
]

pylint.max-branches = 25
pylint.max-statements = 100
pylint.max-args = 6
ignore = [
    "F403", # 'from module import *' used; unable to detect undefined names
    "PLR2004",
    "PLC0415"
]

fixable = ["ALL"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# this is entirely optional, you can remove this if you wish to
[tool.ruff.format]
# use single quotes for strings.
quote-style = "single"

# indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[project.entry-points.'nomad.plugin']
tfsc_general_parser_entry_point = "nomad_tfsc_general.parsers:tfsc_general_parser"
tfsc_general_experiment_experiment_parser_entry_point = "nomad_tfsc_general.parsers:tfsc_general_experiment_experiment_parser"
tfsc_general_package_entry_point = "nomad_tfsc_general.schema_packages:tfsc_general_package"
tfsc_voila_documentation_app_entry_point = "nomad_tfsc_general.apps:tfsc_voila_documentation_app"
tfsc_perseus_sample_search_app_entry_point = "nomad_tfsc_general.apps:perseus_sample_search_app"


[tool.cruft]
# Avoid updating workflow files, this leads to permissions issues
skip = [".github/*"]
